Set a fixed best size for wxMediaCtrl
Previously, wxMediaCtrl would set its best size to the dimensions of the
current video that was loaded.  If a wide video was loaded, this could
cause wxMediaCtrl to take up the entire space, hiding any other widget
that might exist within the same sizer.  Avoid this by setting a
reasonable default size, 480p with 16:9 ratio as most videos are in 16:9
aspect ratio currently.