finally uncovered thee cause of WTF commit SHA-1: 93781f2c67d67d6b286d525c926cb34edc9c2a5c: turns out that for *some* sample applications IsAllowedTraceMask() is invoked *before* the static wxArrayString (implicit) initializer itself, resulting in havoc in the internals of the vector class.
This was uncovered when the code was changed once again, while hunting this issue, into using a local function as the initializer call for the static store. Only once that code was altered to keep the static store INSIDE said function AND convert that function to a GETTER-type function (producing a *reference* to the static global instance), did the crazy assertion failures in debug mode stop, as now the static global store is guaranteed to be initialized at the first use.