debug feeature: (Part 2) improve heap debugging on MSVC using the compiler-provided MSVCRT/CRTDBG plus a bit of juice of our own.
The markers are defined in the external mupdf monolithic library code by way of `wx/debugheap.h`. See the `GerHobbelt/mupdf.git` project. When this library does not define HAVE_MUPDF a default (nil) implementation is used instead.

This is the last part of the modification to have all init-time heap allocations (most of whom are without file+sourceline thanks to using stl classes where we cannot modify the allocator to the debug version) wrapped by a start/end block listing the outer caller source location, which is then used in the leak reporter to provide some decent info about the memory leaks and why they may exist.