GTK: Final fix wxGUIEventLoop::DoYieldFor compatibility when using with CEF
We'll now do Yield differently based on whether we're linking to CEF.
When not using CEF, we maintain the original Yield logic.
When using CEF, we do not replace the global GDK event handler, we only handle GDK events,
and we also do not use gtk_main_iteration() to handle non GDK events, because it will trapped us in and
endless message loop, Pending() will always be true.